+++ modules/aspects/default.nix
@@ -1,0 +1,8 @@
+{ ... }:
+{
+  imports = [
+    ./home-manager-defaults.nix
+    ./nixoa-host.nix
+    ./nixoa-user.nix
+  ];
+}