+++ modules/aspects/nixoa-user.nix
@@ -1,0 +1,4 @@
+{ ... }:
+{
+  den.aspects.nixoaUser.homeManager = import ../_homeManager/profile;
+}