+++ modules/outputs/default.nix
@@ -1,0 +1,7 @@
+{ ... }:
+{
+  imports = [
+    ./apps.nix
+    ./dev-shell.nix
+  ];
+}