+++ modules/topology/default.nix
@@ -1,0 +1,7 @@
+{ ... }:
+{
+  imports = [
+    ./schema.nix
+    ./hosts.nix
+  ];
+}