+++ modules/topology/schema.nix
@@ -1,0 +1,9 @@
+{
+  lib,
+  ...
+}:
+{
+  den.schema.user = {
+    config.classes = lib.mkDefault [ "homeManager" ];
+  };
+}