+++ modules/user.nix
@@ -1,7 +1,0 @@
-{
-  vars,
-  ...
-}:
-{
-  den.aspects.${vars.username}.homeManager = import ./_user/user/home.nix;
-}